+++ package.json
@@ -12,9 +12,9 @@
     "generate:icons": "ts-node src/components/icons/index.ts"
   },
   "dependencies": {
-    "@dnd-kit/core": "^6.3.1",
-    "@dnd-kit/sortable": "^10.0.0",
-    "@dnd-kit/utilities": "^3.2.2",
+    "@-kit/core": "^6.3.1",
+    "@-kit/sortable": "^10.0.0",
+    "@-kit/utilities": "^3.2.2",
     "@hookform/resolvers": "^5.0.1",
     "@radix-ui/react-alert-dialog": "^1.1.6",
     "@radix-ui/react-checkbox": "^1.1.4",
