+++ src/components/Sortable.tsx
@@ -1,0 +1,39 @@
+"use client";
+
+import {
+  closestCenter,
+  DndContext,
+  PointerSensor,
+  useSensor,
+  useSensors,
+} from "@dnd-kit/core";
+import {
+  SortableContext,
+  verticalListSortingStrategy,
+} from "@dnd-kit/sortable";
+import { PropsWithChildren } from "react";
+
+interface IProps {
+  items: any[];
+  onDragEnd: (event: any) => void;
+}
+
+export default function Sortable({
+  items,
+  onDragEnd,
+  children,
+}: PropsWithChildren<IProps>) {
+  const sensors = useSensors(useSensor(PointerSensor));
+
+  return (
+    <DndContext
+      sensors={sensors}
+      collisionDetection={closestCenter}
+      onDragEnd={onDragEnd}
+    >
+      <SortableContext items={items} strategy={verticalListSortingStrategy}>
+        {children}
+      </SortableContext>
+    </DndContext>
+  );
+}
