+++ src/components/common/DateRangePicker.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import Button from "@/components/common/Button/Button";
-import Calendar from "@/components/common/Calender";
 import {
   Popover,
   PopoverContent,
@@ -10,8 +9,13 @@ import {
 import cn from "@/lib/utils";
 import { addDays, format } from "date-fns";
 import { CalendarDays as CalendarIcon } from "lucide-react";
+import dynamic from "next/dynamic";
 import * as React from "react";
 import { DateRange } from "react-day-picker";
+
+const Calendar = dynamic(() => import("@/components/common/Calender"), {
+  ssr: false,
+});
 
 export default function DatePickerWithRange({
   className,
